requirements.txt
aiogram==2.24
python-dotenv==0.21.1
asyncpg==0.27.0
pillow==9.5.0
